+++ pydidas/gui/frames/builders/quick_integration_frame_builder.py
@@ -133,16 +133,28 @@ class QuickIntegrationFrameBuilder:
             parent_widget="tab_plot",
             visible=False,
         )
+        cls._frame.create_check_box(
+            "two_click_selection",
+            "Use 2-click point selection",
+            checked=True,
+            gridPos=(0, 0, 1, 1),
+            parent_widget="input_plot_bc_selection",
+            toolTip=(
+                "The 2-click point selection requires two clicks in the image to "
+                "select a new point: The first click zooms in on the selected point "
+                "while the second click confirms the (finer) selection."
+            ),
+        )
         cls._frame.create_param_widget(
             cls._frame.get_param("overlay_color"),
-            gridPos=(0, 0, 1, 1),
+            gridPos=(-1, 0, 1, 1),
             linebreak=True,
             parent_widget="input_plot_bc_selection",
         )
         cls._frame.add_any_widget(
             "input_beamcenter_points",
             PointPositionTableWidget(cls._frame._widgets["input_plot"]),
-            gridPos=(1, 0, 1, 1),
+            gridPos=(-1, 0, 1, 1),
             parent_widget="input_plot_bc_selection",
         )
         cls._frame.create_any_widget(